// vendor/raw.githubusercontent.com/#mikaelporttila_0c368/rss/refs/tags/1.0.3/src/types/fields/media_rss_fields.ts
export enum MediaRssFields {
  Rating = "media:rating",
  Title = "media:title",
  Description = "media:description",
  Content = "media:content",
  Group = "media:group",
  Keywords = "media:keywords",
  Thumbnails = "media:thumbnails",
  Category = "media:category",
  Hash = "media:hash",
  Player = "media:player",
  Credit = "media:credit",
  Copyright = "media:copyright",
  Text = "media:text",
  Restriction = "media:restriction",
  Community = "media:community",
  Comments = "media:comments",
  Comment = "media:comment",
  Embed = "media:embed",
  Param = "media:param",
  Responses = "media:responses",
  Response = "media:response",
  StarRating = "media:starRating",
  Statistics = "media:statistics",
  BackLinks = "media:backLinks",
  BackLink = "media:backLink",
  Status = "media:status",
  Price = "media:price",
  License = "media:license",
  Subtitle = "media:subTitle",
  PeerLink = "media:peerLink",
  Rights = "media:rights",
  Scenes = "media:scenes",
  Scene = "media:scene",
  Tags = "media:tags",
  PriceValue = "price",
}